Refactoring Domain into Vaadin Project Structure
