+++ ALSpringBackend/src/main/java/com/ardaslegends/data/domain/RegionType.java
@@ -1,4 +1,4 @@
-package com.ardaslegends.alspringbackend.domain;
+package com.ardaslegends.data.domain;
 
 public enum RegionType {
