+++ ALSpringBackend/src/main/java/com/ardaslegends/data/service/SamplePersonService.java
@@ -3,6 +3,8 @@ package com.ardaslegends.data.service;
 import com.ardaslegends.data.entity.SamplePerson;
 import java.util.Optional;
 import java.util.UUID;
+
+import com.ardaslegends.data.repository.SamplePersonRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
